feat(Upload): remove upload state from the Upload context. Instead use a store for it
