+++ package.json
@@ -88,7 +88,8 @@
     "universal-cookie": "^4.0.4",
     "uuid": "^8.3.2",
     "webpack": "^5.37.1",
-    "ws": "^7.4.3"
+    "ws": "^7.4.3",
+    "zustand": "^3.6.5"
   },
   "devDependencies": {
     "@babel/core": "^7.15.0",
